feat(resultsViewTopLoci): Moved files into seperate folder
