+++ src/Analysis/GWASResults/Views/Results/ResultsPheWeb/TopLociTable/downloadTSVFromJSON.js
