+++ src/Analysis/GWASResults/Views/Results/ResultsPheWeb/TopLociTable/filterLociTableData.js
